Restructure passport setup in config/passportSetup
