+++ config/passportSetup.js
@@ -1,0 +1,71 @@
+import passport from 'passport';
+import dotenv from 'dotenv';
+import GoogleStrategy from 'passport-google-oauth20';
+import GitHubStrategy from 'passport-github2';
+import User2 from '../models/User2.js';
+
+dotenv.config();
+
+passport.use(
+  new GoogleStrategy(
+    {
+      clientID: process.env.GOOGLE_CLIENT_ID,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET,
+      callbackURL: '/auth/google/callback',
+    },
+    async (accessToken, refreshToken, profile, cb) => {
+      const newUser = {
+        userId: profile.id,
+        name: profile.displayName,
+        email: profile.emails[0].value,
+        image: profile.photos[0].value,
+      };
+
+      try {
+        let user = await User2.findOne({ userId: profile.id });
+        if (!user) {
+          user = await User2.create(newUser);
+        }
+        cb(null, user);
+      } catch (error) {
+        console.error(error);
+      }
+    },
+  ),
+);
+
+passport.use(
+  new GitHubStrategy(
+    {
+      clientID: process.env.GITHUB_CLIENT_ID,
+      clientSecret: process.env.GITHUB_CLIENT_SECRET,
+      callbackURL: '/auth/github/callback',
+    },
+    async (accessToken, refreshToken, profile, cb) => {
+      const newUser = {
+        userId: profile.id,
+        name: profile.displayName,
+        email: profile.emails[0].value,
+        image: profile.photos[0].value,
+      };
+
+      try {
+        let user = await User2.findOne({ userId: profile.id });
+        if (!user) {
+          user = await User2.create(newUser);
+        }
+        cb(null, user);
+      } catch (error) {
+        console.error(error);
+      }
+    },
+  ),
+);
+
+passport.serializeUser((user, done) => {
+  done(null, user.id);
+});
+
+passport.deserializeUser((userId, done) => {
+  User2.findById(userId, (err, user) => done(err, user));
+});
